JsonSvcEx v0.1.1 Kafka.Helpers

Kafka.Helpers provides helper functions to simplify producing, consuming and interacting with kafka topics

Functions

encode_payload_request(normalized_payload)

encode_payload_request(map) :: String.t

creates a json payload from an incoming map

fetch_handler_ids(handlers)

fetch_handler_ids([reference]) :: [atom]

fetches uniq worker ids from a collection of kafka worker module names via kafka_meta/0

fetch_worker_ids(topics)

fetch_worker_ids([reference]) :: atom | [atom]

collects normalized payloads of uniq consumers, handlers and expected batch sizes

format_time(datetime)

format_time(DateTime.t) :: String.t

formate a timestamp to a iso formated datetime

Examples

iex> timestamp = "2017-03-19T02:48:15.814147Z"
iex> {:ok, time, _} = DateTime.from_iso8601(timestamp)
iex> Kafka.Helpers.format_time(time)
"2017-03-19T02:48:15.814147Z"

now()

now :: DateTime.t

generate a utc standardized timestamp

now_string()

genreates an iso8601 now timestamp

payload_normalizer(payload, timestamp, uuid)

payload_normalizer(map, String.t, String.t) :: map

Normalizes an outgoing payload Examples

iex> payload = %{hello: :world}
iex> time = "2017-03-19T02:48:15.814147Z"
iex> uuid = "33c5aa7e-3f35-47bc-883b-33ea0ace89f0"
iex> Kafka.Helpers.payload_normalizer(payload, time, uuid)
%{meta: %{
  requested_at: "2017-03-19T02:48:15.814147Z",
  transaction_id: "33c5aa7e-3f35-47bc-883b-33ea0ace89f0"},
request_body: %{hello: :world}}

select_random_partition(partitions)

select_random_partition([integer]) :: integer

selects a random value from a set of partions

Examples

iex> partitions = [1]
iex> Kafka.Helpers.select_random_partition(partitions)
1

start_consumers(topics, handler \\ &KafkaEx.fetch/3)

start_consumers([String.t], (... -> any)) :: {:ok, [pid]}

starts a collection of consumer workers defaulting to KafkaEx.fetch/3 for the message handler

validate_uuid(uuid)

validate_uuid(String.t) ::
  {:ok, String.t} |
  {:error, :invalid_uuid}

validates a string as a uuid

Examples

iex> Kafka.Helpers.validate_uuid("33c5aa7e-3f35-47bc-883b-33ea0ace89f0")
{:ok, "33c5aa7e-3f35-47bc-883b-33ea0ace89f0"}

iex> Kafka.Helpers.validate_uuid("some_string")
{:error, :invalid_uuid}