JsonSvcEx v0.1.0 KafkaHandlers.Account.RequestCreate

KafkaHandlers for account controller Publishes to kafka topic using worker defined by @worker_id a random partition is selected per message TODO: include pattern matching for versioned topics

Functions

kafka_meta()

publish(account_payload)

publish(map) :: map

publish(account_payload, list)

publish(map, %{timestamp: DateTime}) :: map