JsonSvcEx v0.1.1 KafkaHandlers.Workers

Kafka Workers - create workers with atoms for handling different controller streams @workers list of worker modules that contain interface kafka_meta

Functions

create_workers(handler \\ &KafkaEx.create_worker/1)

This function creates workers injected from the array @workers workers must impliment .kafka_meta[:worker_id]