JsonSvcEx v0.1.1 Model.Account.Request

Account request object

Examples

iex> test_model = %Model.Account.Request{
...> first_name: "Joe",
...> last_name: "Bob",
...> email: "joe@bob.com"}
iex> Model.Account.Request.validate!(test_model)
{:ok, :valid}

Functions

validate!(request)

validate!(%{first_name: String.t, last_name: String.t, email: String.t}) :: {:ok, :valid}

This function validates an request object required keys are validated if all required keys are not present an AppError is raised. Honeybadger error is ignored. http status code is set to 400 for an invalid input error