JsonSvcEx v0.1.1 Response.Json

Poison Encoder for handling json response bodies with status code sets default values for body=%{} and status=200 supports skipping one or both extra params

Functions

fail(conn)

fail(Plug.Conn) :: no_return

This function sends a failure message to a Plug.Conn connected client

Parameters

conn - expects Plug.Conn

message: Map or String error message to report to client

http_code: Integer http status code to send to client

fail(conn, payload)

fail(Plug.Conn, %{message: String.t, http_code: pos_integer}) :: no_return

parse(conn)

parse(Plug.Conn) :: Map.t

parse(conn, list)

parse(Plug.Conn, [{:type, any}]) :: any

This function parses an incoming json response body

Parameters

conn - expectes Plug.Conn

type - passed in type to render the incoming payload as

render(conn)

render(Plug.Conn) :: no_return

This function renders a result back to an incoming connection

Parameters

conn - expectes Plug.Conn

body - response body Map or String that is parsed into Json

status - http status code Integer

render(conn, list)

render(Plug.Conn, [{:body, Map.t}]) :: no_return

render(Plug.Conn, [{:status, pos_integer}]) :: no_return

render(Plug.Conn, body: Map.t, status: pos_integer) :: no_return